fix(mypage): 뱃지 target 타입을 서버 double에 맞춤
GET /users/me/badges가 통째로 실패하고 있었다. 서버는 target을 double로
주는데(OpenAPI BadgeResponse: progress int32 / target double) DTO가 Int?라
"target":1.0에서 역직렬화 예외가 났다 — 부분 degrade가 아니라 AppError.Unknown.

매퍼 테스트가 이를 가리고 있었다. 픽스처의 target이 정수("target":10)라
통과했고 실서버만 실패했다. 픽스처를 실제 응답 형태(10.0)로 고쳐 같은 함정을
테스트가 잡게 했다.

도메인 progressTarget(Int?)은 그대로 두고 데이터 계층에서만 변환해
UI·iOS 미러는 건드리지 않았다 (규칙 §6).

검증: 로컬 백엔드 실응답으로 회원 조회 12건 전부 파싱 확인

+++ shared/src/commonMain/kotlin/org/sesacteamproject/passmate/user/data/dto/BadgesResponse.kt
@@ -19,6 +19,7 @@ data class BadgesResponse(
         val achieved: Boolean = false,
         val achievedAt: String? = null,
         val progress: Int? = null,
-        val target: Int? = null
+        // 서버는 target만 double로 준다 (OpenAPI BadgeResponse: progress int32 / target double)
+        val target: Double? = null
     )
 }
